backend/requirements.txt
fastapi
uvicorn
gunicorn
python-multipart
pandas
openpyxl
google-genai
pydantic
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
